Add script for analyzing experiment types and resolution
